requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9  # For file uploads
opencv-python==4.10.0.84
numpy==1.26.4
openai==1.51.0
pydantic==2.9.2
python-dotenv==1.0.1
boto3==1.35.24
pillow-heif==0.22.0
rawpy==0.24.0
pillow==11.1.0